usb: f_fs: Avoid invalid pointer access in ffs_fs_get_tree
Consider a case where ffs_data_new is getting called from
ffs_fs_get_tree and ffs_data_new returns error pointer since
the function (e.g. adb) is already mounted. The driver is only
checking for the NULL return value for ffs but in this case it
will not be NULL, which will fail the check and driver will go
ahead accessing invalid pointer which can lead to inconsistencies.

Fix this by having NULL as well as error pointer check for ffs.

Change-Id: I50940b0524ae9d6dcf1fa4593d547d1cac66d3b2
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>