drivers: input: sec_ts: Move input open/close control into kernel
Background: FOD isn't working when the ambient display
shows up on an off screen because the input open/close
status was not updated by power HAL, and the low-power
state was not updated.

To address this, move input open/close control into
kernel, and the following implementations are made:
 - Perform input open/close whenever an FB event is
   notified to ensure that the low-power state is
   always up-to-date.
 - If the FB event is FB_BLANK_POWERDOWN, close the
   input.
 - If the FB event is anything else, swap the input
   status, as there is no case where a new FB event
   is notified but no input status change is needed.
 - Set ts->input_close on boot, to open the input
   when boot is finished.

These changes correct the input status to open when
the ambient display shows up and resolve the FOD issue.
Additionally, our power HAL no longer needs to handle
input open/close.

Change-Id: I16a5e9dcccc0e06ee9eff55b52af39793c7261cc